flattens an AFM image using a plane fit this flattens only z, not z.nm, so scaling factor would need to be applied for z.nm

flattens an AFM image using a plane fit this flattens only z, not z.nm, so scaling factor would need to be applied for z.nm

AFM.flatten(obj)

Arguments

obj
AFMdata object

Value

flattened matrix with AFM image

Author

thomasgredig

Examples

d = AFM.read(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
d2 = AFM.flatten(d)
#> Error in paste(AFMcopy@history, "AFM.flatten();"): trying to get slot "history" from an object (class "data.frame") that is not an S4 object