Height-Height Correlation Function

Computes g(r) correlation function for an AFM image with height information. The height information must be encoded in a BMP file with a linear scale Publication: http://iopscience.iop.org/article/10.1088/1742-6596/417/1/012069 Title: Height-Height Correlation Function to Determine Grain Size in Iron Phthalocyanine Thin Films Authors: Thomas Gredig, Evan A. Silverstein, Matthew P Byrne Journal: J of Phys: Conf. Ser. Vol 417, p. 012069 (2013).

AFM.hhcf(
  obj,
  no = 1,
  degRes = 100,
  numIterations = 1000,
  addFit = TRUE,
  dataOnly = FALSE,
  r.percentage = 80,
  xi.percentage = 70,
  verbose = FALSE
)

Arguments

obj

AFMdata object

no

Channel number

degRes

resolution of angle, the higher the more, should be >100, 1000 is also good

numIterations

Number of iterations (must be > 1000), but 1e6 recommended

addFit

if TRUE a fit is added to the data

dataOnly

if TRUE only return data frame, otherwise returns a graph

r.percentage

a number from 10 to 100 representing the distance to compute, since the image is square, there are not as many points that are separate by the full lenght, 80 is a good value, if there is no fit, the value can be reduced to 70 or 60.

xi.percentage

a number from 10 to 100 representing where correlation length could be found from maximum (used for fitting)

verbose

output time if TRUE

Value

graph or data frame with g(r) and $num indicating number of computations used for r

Author

Thomas Gredig

Examples

filename = AFM.getSampleImages(type='tiff')
a = AFM.import(filename)
a = AFM.flatten(a)
r = AFM.hhcf(a, dataOnly = TRUE)
head(r)
#>        r.nm          g num
#> 1  9.803922 0.08883411 998
#> 2 19.607843 0.21919080 996
#> 3 29.411765 0.36936237 993
#> 4 39.215686 0.47572459 985
#> 5 49.019608 0.56649596 977
#> 6 58.823529 0.63815281 971