AFM image class

A S4 class to store and manipulate images from Atomic Force Microscopes. It contains the names of channels (exact naming depends on the type/instrument) The class supports multiple images; it can also contain additional information such as lines, etc.

AFMdata(
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history,
  description = "",
  fullfilename
)

AFMdata(
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history,
  description = "",
  fullfilename
)

Arguments

data
list with objects ($z is a list with images)

x.conv
conversion factor from pixels to nm

y.conv
conversion factor from pixels to nm

x.pixels
number of pixels in x-direction

y.pixels
number of pixels in y-direction

z.conv
(not used)

z.units
vector with units for $z (deg, m)

channel
vector with names of channels

instrument
name of instrument (Park, Cypher, NanoSurf, Veeco)

history
history of file changes

description
AFM image description or note

fullfilename
name of file

Slots

data

list with objects ($z is a list with images)

x.conv

conversion factor from pixels to nm

y.conv

conversion factor from pixels to nm

x.pixels

number of pixels in x-direction

y.pixels

number of pixels in y-direction

x.nm

length of image

y.nm

height of image

z.conv

(not used)

z.units

vector with units for $z (deg, m)

channel

vector with names of channels

instrument

name of instrument (Park, Cypher, NanoSurf, Veeco)

history

history of file changes

description

AFM image description or note

fullfilename

name of file