loads header of NID file and separates into items first item in list has the titles of the others

loads header of NID file and separates into items first item in list has the titles of the others

read.NID_headerItems(filename)

Arguments

filename
filename including path

Value

list

Examples

filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
read.NID_headerItems(filename)
#> [[1]]
#>  [1] "HEADERS"                             "[DataSet]"                          
#>  [3] "[DataSet-Info]"                      "[DataSet\\DataSetInfos]"            
#>  [5] "[DataSet\\DataSetInfos\\Scan]"       "[DataSet\\DataSetInfos\\Feedback]"  
#>  [7] "[DataSet\\DataSetInfos\\Global]"     "[DataSet\\DataSetInfos\\Module]"    
#>  [9] "[DataSet\\Calibration]"              "[DataSet\\Calibration\\Scanhead]"   
#> [11] "[DataSet\\Calibration\\Cantilever]"  "[DataSet\\Parameters]"              
#> [13] "[DataSet\\Parameters\\Approach]"     "[DataSet\\Parameters\\ZFeedback]"   
#> [15] "[DataSet\\Parameters\\Lithography]"  "[DataSet\\Parameters\\Imaging]"     
#> [17] "[DataSet\\Parameters\\SignalIO]"     "[DataSet\\Parameters\\Spectroscopy]"
#> [19] "[DataSet\\Parameters\\SPMSystem]"    "[DataSet-0:1]"                      
#> [21] "[DataSet-0:2]"                       "[DataSet-1:1]"                      
#> [23] "[DataSet-1:2]"                       "[SetView]"                          
#> [25] "[SetView-View0]"                     "[SetView-View1]"                    
#> [27] "[SetView-View2]"                     "[SetView-View3]"                    
#> 
#> [[2]]
#>  [1] "[DataSet]"              "Version=2"              "GroupCount=2"          
#>  [4] "Gr0-Name=Scan forward"  "Gr0-ID=0"               "Gr0-Count=21"          
#>  [7] "Gr1-Name=Scan backward" "Gr1-ID=0"               "Gr1-Count=21"          
#> [10] "Gr0-Ch1=DataSet-0:1"    "Gr0-Ch2=DataSet-0:2"    "Gr1-Ch1=DataSet-1:1"   
#> [13] "Gr1-Ch2=DataSet-1:2"    "SetInfos=DataSet-Info"  "InfoSetCount=3"        
#> [16] "InfoSet0=DataSetInfos"  "InfoSet1=Calibration"   "InfoSet2=Parameters"   
#> 
#> [[3]]
#>  [1] "[DataSet-Info]"                   "-- Scan --=--------"             
#>  [3] "Image size=10\xb5m"               "Scan direction=Up"               
#>  [5] "Time/Line=1 s"                    "Points=128 "                     
#>  [7] "Lines=128 "                       "X-Slope=0 \xb0"                  
#>  [9] "Y-Slope=0 \xb0"                   "Rotation=0 \xb0"                 
#> [11] "X-Pos=0 m"                        "Y-Pos=0 m"                       
#> [13] "Z-Plane=0 m"                      "Const. Height-Mode=Disabled"     
#> [15] "Date=01-03-2016"                  "Time=18:36:58"                   
#> [17] "-- Feedback --=--------"          "Setpoint=50 %"                   
#> [19] "P-Gain=10000 "                    "I-Gain=1000 "                    
#> [21] "D-Gain=0 "                        "P-Gain2=0 "                      
#> [23] "I-Gain2=1000 "                    "D-Gain2=0 "                      
#> [25] "Tip voltage=0 V"                  "Feedback mode=Free"              
#> [27] "Feedback algo.=Adaptive PID"      "Vibration freq.=187.396kHz"      
#> [29] "Vibration ampl.=100mV"            "Excitation ampl.=268mV"          
#> [31] "Error range=5 V"                  "Ampl. Ctrl. mode=Free Amp."      
#> [33] "-- Global --=--------"            "Measurement environment=Air"     
#> [35] "Op. mode=Dynamic Force"           "Cantilever type=ACL-A"           
#> [37] "Head type=EZ2-AFM"                "Scan head=Uncal_EZ2-AFM 110u.hed"
#> [39] "Laser working point=0.0%"         "Deflection offset=0.0%"          
#> [41] "Software ver.=3.3.0.3"            "Firmware ver.=3.3.0.0"           
#> [43] "Controller S/N=023-08-554"        "-- Module --=--------"           
#> [45] "Controller Board=2"               "AFM Basic Module=3"              
#> [47] "AFM Dynamic Module=2"             "USB Module=3"                    
#> [49] "Scripting Interface=1"            "Spectroscopy Module=1"           
#> [51] "Lithography Module=1"            
#> 
#> [[4]]
#> [1] "[DataSet\\DataSetInfos]" "SubSectionCount=4"      
#> [3] "SubSection0=Scan"        "SubSection1=Feedback"   
#> [5] "SubSection2=Global"      "SubSection3=Module"     
#> 
#> [[5]]
#>  [1] "[DataSet\\DataSetInfos\\Scan]" "SubSectionCount=0"            
#>  [3] "Image size=10\xb5m"            "Scan direction=Up"            
#>  [5] "Time/Line=1 s"                 "Points=128 "                  
#>  [7] "Lines=128 "                    "X-Slope=0 \xb0"               
#>  [9] "Y-Slope=0 \xb0"                "Rotation=0 \xb0"              
#> [11] "X-Pos=0 m"                     "Y-Pos=0 m"                    
#> [13] "Z-Plane=0 m"                   "Const. Height-Mode=Disabled"  
#> [15] "Date=01-03-2016"               "Time=18:36:58"                
#> 
#> [[6]]
#>  [1] "[DataSet\\DataSetInfos\\Feedback]" "SubSectionCount=0"                
#>  [3] "Setpoint=50 %"                     "P-Gain=10000 "                    
#>  [5] "I-Gain=1000 "                      "D-Gain=0 "                        
#>  [7] "P-Gain2=0 "                        "I-Gain2=1000 "                    
#>  [9] "D-Gain2=0 "                        "Tip voltage=0 V"                  
#> [11] "Feedback mode=Free"                "Feedback algo.=Adaptive PID"      
#> [13] "Vibration freq.=187.396kHz"        "Vibration ampl.=100mV"            
#> [15] "Excitation ampl.=268mV"            "Error range=5 V"                  
#> [17] "Ampl. Ctrl. mode=Free Amp."       
#> 
#> [[7]]
#>  [1] "[DataSet\\DataSetInfos\\Global]"  "SubSectionCount=0"               
#>  [3] "Measurement environment=Air"      "Op. mode=Dynamic Force"          
#>  [5] "Cantilever type=ACL-A"            "Head type=EZ2-AFM"               
#>  [7] "Scan head=Uncal_EZ2-AFM 110u.hed" "Laser working point=0.0%"        
#>  [9] "Deflection offset=0.0%"           "Software ver.=3.3.0.3"           
#> [11] "Firmware ver.=3.3.0.0"            "Controller S/N=023-08-554"       
#> 
#> [[8]]
#> [1] "[DataSet\\DataSetInfos\\Module]" "SubSectionCount=0"              
#> [3] "Controller Board=2"              "AFM Basic Module=3"             
#> [5] "AFM Dynamic Module=2"            "USB Module=3"                   
#> [7] "Scripting Interface=1"           "Spectroscopy Module=1"          
#> [9] "Lithography Module=1"           
#> 
#> [[9]]
#> [1] "[DataSet\\Calibration]" "SubSectionCount=2"      "SubSection0=Scanhead"  
#> [4] "SubSection1=Cantilever"
#> 
#> [[10]]
#>  [1] "[DataSet\\Calibration\\Scanhead]"                                             
#>  [2] "SubSectionCount=0"                                                            
#>  [3] "Version=5"                                                                    
#>  [4] "HeadTyp=EZ2-AFM"                                                              
#>  [5] "SerialNo=uncal 110u"                                                          
#>  [6] "RevisionNo=2"                                                                 
#>  [7] "ScanFit=Pomfit"                                                               
#>  [8] "CtrlInPol=Positiv"                                                            
#>  [9] "AnalogOut2Pol=Negativ"                                                        
#> [10] "AnalogOut3Pol=Negativ"                                                        
#> [11] "ZAxisPol=Negativ"                                                             
#> [12] "ZAxisAnalogOut2Pol=Negativ"                                                   
#> [13] "ZAxisAnalogOut3Pol=Negativ"                                                   
#> [14] "Ch0Corr=None"                                                                 
#> [15] "SetPointCheck=None"                                                           
#> [16] "InCount=19"                                                                   
#> [17] "In0=0,Channel0,Amplitude,V,2.5,0"                                             
#> [18] "In1=1,Channel1,Z-Axis,m,1e-005,0"                                             
#> [19] "In2=2,Channel2,Amplitude,V,2.5,0"                                             
#> [20] "In3=3,Channel3,Phase,\xb0,90,0"                                               
#> [21] "In4=4,Channel4,Tip current,A,0.0001,0"                                        
#> [22] "In5=5,TipSignalDC,Deflection,m,8e-006,0"                                      
#> [23] "In6=6,TipSignalAC,Amplitude,V,10,0"                                           
#> [24] "In7=7,TipPhase,Phase,\xb0,90,0"                                               
#> [25] "In8=8,Lever Current,Tip current,A,0.0001,0"                                   
#> [26] "In9=9,User Input0,User input 1,V,10,0"                                        
#> [27] "In10=10,User Input1,User input 2,V,10,0"                                      
#> [28] "In11=11,HeadSig2,Not defined 1,V,10,0"                                        
#> [29] "In12=12,Channel5,User input 1,V,10,0"                                         
#> [30] "In13=13,Channel6,User input 2,V,10,0"                                         
#> [31] "In14=16,TipSignalDC1,Deflection,V,10,0"                                       
#> [32] "In15=17,TipSignalDC2,Deflection,N,1,0"                                        
#> [33] "In16=18,SensorSignal1,SensorSignal,W,0.005,0"                                 
#> [34] "In17=19,SensorSignal2,SensorSignal,A,0.0005,0"                                
#> [35] "In18=20,Channel7,Z-Axis,m,1e-005,0"                                           
#> [36] "OutCount=9"                                                                   
#> [37] "Out0=0,ScanAxis0,X-Axis,m,5.5e-005,0"                                         
#> [38] "Out1=1,ScanAxis1,Y-Axis,m,5.5e-005,0"                                         
#> [39] "Out2=2,ScanAxis2,Z-Axis,m,1e-005,0"                                           
#> [40] "Out3=4,AnalogOut0,Tip voltage,V,10,0"                                         
#> [41] "Out4=5,AnalogOut1,Not defined,V,10,0"                                         
#> [42] "Out5=6,AnalogOut2,User Output 1,V,10,0"                                       
#> [43] "Out6=7,AnalogOut3,User Output 2,V,10,0"                                       
#> [44] "Out7=8,DriveAmp,Excitation amplitude,V,5.3,0"                                 
#> [45] "Out8=9,MixedOut4,User Output C,V,10,0"                                        
#> [46] "ScanCorrCount=1"                                                              
#> [47] "ScanCorrA=0"                                                                  
#> [48] "ScanCorrB=0"                                                                  
#> [49] "ScanCorrC=0"                                                                  
#> [50] "ScanCorrD=0"                                                                  
#> [51] "ScanCorrE=0"                                                                  
#> [52] "ScanCorrF=0"                                                                  
#> [53] "ScanCorrG=0"                                                                  
#> [54] "ScanCorrH=0"                                                                  
#> [55] "ScanCorrI=0"                                                                  
#> [56] "ScanCorrJ=0"                                                                  
#> [57] "ScanCorrK=0"                                                                  
#> [58] "ScanCorrL=0"                                                                  
#> [59] "ScanCorrM=0"                                                                  
#> [60] "ScanCorrN=0"                                                                  
#> [61] "ScanCorrO=0"                                                                  
#> [62] "ScanCorrTableRot=0"                                                           
#> [63] "ScanCorrRotX=0"                                                               
#> [64] "ScanCorrRotY=0"                                                               
#> [65] "ScanScaleZ=3276.8"                                                            
#> [66] "ScanScaleXY=256"                                                              
#> [67] "ZRangeFit=None"                                                               
#> [68] "ZRangeCorrCount=1"                                                            
#> [69] "ZRangeCorrA=1"                                                                
#> [70] "ZRangeCorrB=0"                                                                
#> [71] "ZRangeCorrC=0"                                                                
#> [72] "ZRangeCorrD=0"                                                                
#> [73] "ApproachStartPos=0"                                                           
#> [74] "MaxApproachSpeed=1"                                                           
#> [75] "ManualMoveSpeed=1"                                                            
#> [76] "DefaultDynRefAmp=0.25"                                                        
#> [77] "DefaultDynRefAmpLiquid=0.2"                                                   
#> [78] "DefaultForceModRefAmp=1"                                                      
#> [79] "DetectorABGainAC=0.5"                                                         
#> [80] "CompDcWithZ=0"                                                                
#> [81] "VideoSupport=1"                                                               
#> [82] "SensorStatusMode=1"                                                           
#> [83] "SensorStatusPol=Negativ"                                                      
#> [84] "SensorStatusMeterRange=0.0866142;0.102362;0.102362;0.811024;0.811024;0.913386"
#> [85] "LaserWavelength=0"                                                            
#> 
#> [[11]]
#>  [1] "[DataSet\\Calibration\\Cantilever]"         
#>  [2] "SubSectionCount=0"                          
#>  [3] "Version=2"                                  
#>  [4] "GUID={ABB75273-9543-431a-B681-C79B533DD9E6}"
#>  [5] "Name=ACL-A"                                 
#>  [6] "Manufacturer=AppNano"                       
#>  [7] "PropCount=10"                               
#>  [8] "Prop0=D[48]*[N/m]"                          
#>  [9] "Prop1=D[190000]*[Hz]"                       
#> [10] "Prop2=D[0.000225]*[m]"                      
#> [11] "Prop3=D[4.3e-005]*[m]"                      
#> [12] "Prop4=D[500]*[]"                            
#> [13] "Prop5=D[65000]*[Hz]"                        
#> [14] "Prop6=D[5]*[]"                              
#> [15] "Prop7=D[0]*[]"                              
#> [16] "Prop8=D[18]*[\xb0]"                         
#> [17] "Prop9=D[6e-009]*[m]"                        
#> 
#> [[12]]
#> [1] "[DataSet\\Parameters]"    "SubSectionCount=7"       
#> [3] "SubSection0=Approach"     "SubSection1=ZFeedback"   
#> [5] "SubSection2=Lithography"  "SubSection3=Imaging"     
#> [7] "SubSection4=SignalIO"     "SubSection5=Spectroscopy"
#> [9] "SubSection6=SPMSystem"   
#> 
#> [[13]]
#>  [1] "[DataSet\\Parameters\\Approach]"    "SubSectionCount=0"                 
#>  [3] "ApproachSteps=L[65000]*[]"          "WithdrawSteps=L[1000]*[]"          
#>  [5] "AutoPeriode=D[0.02]*[s]"            "MovePeriode=D[0.02]*[s]"           
#>  [7] "StepSlope=L[6]*[]"                  "AutoStartScan=B[0]*[]"             
#>  [9] "AutoReloadParameter=B[0]*[]"        "ApproachSpeed=D[0.0801099]*[%]"    
#> [11] "ManualSpeed=D[1]*[]"                "AFMApproachMode=L[0]*[]"           
#> [13] "AFMStepByStepSpeed=D[1]*[]"         "AFMStepByStepSlope=D[5e-006]*[m/s]"
#> 
#> [[14]]
#>  [1] "[DataSet\\Parameters\\ZFeedback]"   "SubSectionCount=0"                 
#>  [3] "PGain=L[10000]*[]"                  "IGain=L[1000]*[]"                  
#>  [5] "DGain=L[0]*[]"                      "PGain2=L[0]*[]"                    
#>  [7] "IGain2=L[1000]*[]"                  "DGain2=L[0]*[]"                    
#>  [9] "SetPoint=D[50]*[%]"                 "SetPointForce=D[1.99713e-008]*[N]" 
#> [11] "SetPointDC=D[1e-009]*[m]"           "SetPointAmp=D[50]*[%]"             
#> [13] "SetPointPhase=D[1]*[\xb0]"          "SetPointForceUnitMode=L[2]*[]"     
#> [15] "SetPointForce_V=D[0.273]*[]"        "SetPointForce_m=D[0.0001]*[]"      
#> [17] "SetPointForce_N=D[1.99713e-008]*[]" "ChGainRegError=L[0]*[]"            
#> [19] "FeedbackMode=L[0]*[]"               "Algorithm=L[1]*[]"                 
#> [21] "RefFrq=D[187396]*[Hz]"              "RefAmplitude=D[0.225962]*[V]"      
#> [23] "RefPhase=D[112.088]*[\xb0]"         "CoarseSweepStartFrq=D[133000]*[]"  
#> [25] "CoarseSweepEndFrq=D[247000]*[]"     "CoarseSweepStepFrq=D[100.002]*[]"  
#> [27] "FineSweepStartFrq=D[186626]*[]"     "FineSweepEndFrq=D[188125]*[]"      
#> [29] "FineSweepStepFrq=D[10]*[]"          "ManualSweepStartFrq=D[186455]*[]"  
#> [31] "ManualSweepEndFrq=D[187955]*[]"     "ManualSweepStepFrq=D[5.85938]*[]"  
#> [33] "AutoCalFrqPeak=B[1]*[]"             "AutoCalPhase=B[1]*[]"              
#> [35] "DynamicAmp=D[0.100253]*[V]"         "CalRefPlotGraph=B[0]*[]"           
#> [37] "ForceModFrq=D[20000]*[Hz]"          "ForceModAmp=D[4]*[V]"              
#> [39] "AutoFrqSweepRange=B[43]*[]"         "AmpCtrlMode=L[2]*[]"               
#> [41] "FrqPeakReduceAir=D[20]*[%]"         "FrqPeakUpperSideAir=B[1]*[]"       
#> [43] "FrqPeakReduceLiquid=D[5]*[%]"       "FrqPeakUpperSideLiquid=B[0]*[]"    
#> [45] "RegMoveMode=L[0]*[]"                "DynamicMeasureBW=L[6]*[]"          
#> [47] "RegOutputSel=L[0]*[]"               "KPFMModFrq=D[100]*[Hz]"            
#> [49] "KPFMModAmp=D[0]*[V]"                "KPFMModPhase=D[0]*[\xb0]"          
#> [51] "KPFMMeasureBWSel=L[0]*[]"           "KPFMCtrlEnabled=B[0]*[]"           
#> [53] "KPFMSetPoint=D[0]*[V]"              "KPFMPGain=D[0]*[]"                 
#> [55] "KPFMIGain=D[0]*[]"                 
#> 
#> [[15]]
#>  [1] "[DataSet\\Parameters\\Lithography]" "SubSectionCount=0"                 
#>  [3] "ModuleLevel=L[1]*[]"                "OpMode=L[2]*[]"                    
#>  [5] "InactivePenMode=L[0]*[]"            "AutoChartSettings=B[1]*[]"         
#>  [7] "AutoCapture=B[1]*[]"                "XYMoveSpeed=D[1e-005]*[m/s]"       
#>  [9] "ZMoveSpeed=D[1e-006]*[m/s]"         "LiftTipAbsZPos=D[0]*[m]"           
#> 
#> [[16]]
#>  [1] "[DataSet\\Parameters\\Imaging]"        
#>  [2] "SubSectionCount=0"                     
#>  [3] "ModuleLevel=L[1]*[]"                   
#>  [4] "ScanRange=V[1e-005,1e-005,0]*[m,m,m]"  
#>  [5] "ScanOffset=V[0,0,0]*[m,m,m]"           
#>  [6] "ScanRotation=V[0,0,0]*[\xb0,\xb0,\xb0]"
#>  [7] "ScanTime=V[1.00051,1.00051,0]*[s,s,s]" 
#>  [8] "Datapoints=V[128,128,0]*[,,]"          
#>  [9] "MeasureMask=L[3]*[]"                   
#> [10] "ScanMode=L[0]*[]"                      
#> [11] "OverScanSize=D[0]*[%]"                 
#> [12] "FirstScanlineRep=L[1]*[]"              
#> [13] "SyncOutMode=L[0]*[]"                   
#> [14] "LineMode=L[0]*[]"                      
#> [15] "RelTipPos=D[3.2795e-007]*[m]"          
#> [16] "AutoSlopeCorrection=B[0]*[]"           
#> [17] "AutoDeleteBuffer=B[1]*[]"              
#> [18] "AutoChartSettings=B[1]*[]"             
#> [19] "AutoCapture=B[1]*[]"                   
#> 
#> [[17]]
#>  [1] "[DataSet\\Parameters\\SignalIO]" "SubSectionCount=0"              
#>  [3] "UserDAC0=D[0]*[V]"               "UserDAC1=D[0]*[V]"              
#>  [5] "EnableUserADC0=B[0]*[]"          "EnableUserADC1=B[0]*[]"         
#>  [7] "TipSignalMode=L[0]*[]"           "ExcitationMode=L[0]*[]"         
#>  [9] "User0CtrlMode=L[0]*[]"           "User0InputPol=L[0]*[]"          
#> [11] "User0OutputFlag=L[0]*[]"         "User0SetPoint=D[0]*[V]"         
#> [13] "User0IGain=L[1000]*[]"           "MonitorOut0=L[0]*[]"            
#> [15] "MonitorOut1=L[0]*[]"            
#> 
#> [[18]]
#>  [1] "[DataSet\\Parameters\\Spectroscopy]" 
#>  [2] "SubSectionCount=0"                   
#>  [3] "ModuleLevel=L[1]*[]"                 
#>  [4] "SpecPoints=L[1]*[]"                  
#>  [5] "OutputType=L[0]*[]"                  
#>  [6] "Repetition=L[1]*[]"                  
#>  [7] "RepetitionMode=L[0]*[]"              
#>  [8] "FwModTime=D[1]*[s]"                  
#>  [9] "BwModTime=D[1]*[s]"                  
#> [10] "ModRelValue=B[1]*[]"                 
#> [11] "FwModDatapoints=L[1024]*[]"          
#> [12] "BwModDatapoints=L[1024]*[]"          
#> [13] "SyncOutMode=L[0]*[]"                 
#> [14] "FeedbackActive=B[0]*[]"              
#> [15] "AutoChartSettings=B[1]*[]"           
#> [16] "AutoCapture=B[1]*[]"                 
#> [17] "SpecEndMode=L[1]*[]"                 
#> [18] "FwModRange=D[9.45455e-007]*[m]"      
#> [19] "BwModRange=D[-9.45455e-007]*[m]"     
#> [20] "XYMoveSpeed=D[2e-005]*[m/s]"         
#> [21] "StartOffset=D[-5.72715e-007]*[m]"    
#> [22] "StartOffsetMoveSpeed=D[1e-005]*[m/s]"
#> [23] "FwMoveSpeed=D[9.45455e-007]*[m/s]"   
#> [24] "BwMoveSpeed=D[9.45455e-007]*[m/s]"   
#> [25] "FwSamplingRate=D[1024]*[Hz]"         
#> [26] "BwSamplingRate=D[1024]*[Hz]"         
#> [27] "FwModMode=L[0]*[]"                   
#> [28] "BwModMode=L[0]*[]"                   
#> [29] "FwModStopValue=D[0]*[V]"             
#> [30] "BwModStopValue=D[0]*[V]"             
#> [31] "FwModStopMode=L[1]*[]"               
#> [32] "BwModStopMode=L[1]*[]"               
#> [33] "FwPauseTime=D[0]*[s]"                
#> [34] "BwPauseTime=D[0]*[s]"                
#> [35] "FwPauseDatapoints=L[0]*[]"           
#> [36] "BwPauseDatapoints=L[0]*[]"           
#> [37] "FwPauseMode=L[0]*[]"                 
#> [38] "BwPauseMode=L[0]*[]"                 
#> [39] "AutoRecalibrateProbe=B[0]*[]"        
#> 
#> [[19]]
#>  [1] "[DataSet\\Parameters\\SPMSystem]"                               
#>  [2] "SubSectionCount=0"                                              
#>  [3] "AOut0=D[0]*[V]"                                                 
#>  [4] "AOut1=D[0]*[V]"                                                 
#>  [5] "SPMMode=L[3]*[]"                                                
#>  [6] "VideoSource=L[0]*[]"                                            
#>  [7] "VideoTopLight=D[100]*[%]"                                       
#>  [8] "VideoSideLight=D[100]*[%]"                                      
#>  [9] "VideoTopBrightness=D[80]*[]"                                    
#> [10] "VideoSideBrightness=D[80]*[]"                                   
#> [11] "VideoTopContrast=D[70]*[]"                                      
#> [12] "VideoSideContrast=D[70]*[]"                                     
#> [13] "FlexVideoTopExpTime=L[60]*[]"                                   
#> [14] "FlexVideoSideExpTime=L[60]*[]"                                  
#> [15] "FlexVideoTopGain=L[8]*[]"                                       
#> [16] "FlexVideoSideGain=L[8]*[]"                                      
#> [17] "FlexVideoTopADCLevel=L[2]*[]"                                   
#> [18] "FlexVideoSideADCLevel=L[2]*[]"                                  
#> [19] "FlexVideoTopRatioR=D[1]*[]"                                     
#> [20] "FlexVideoSideRatioR=D[1]*[]"                                    
#> [21] "FlexVideoTopRatioG=D[1]*[]"                                     
#> [22] "FlexVideoSideRatioG=D[1]*[]"                                    
#> [23] "FlexVideoTopRatioB=D[1]*[]"                                     
#> [24] "FlexVideoSideRatioB=D[1]*[]"                                    
#> [25] "FlexVideoTopRatioGamma=D[1]*[]"                                 
#> [26] "FlexVideoSideRatioGamma=D[1]*[]"                                
#> [27] "FlexVideoTopAutoLevels=B[1]*[]"                                 
#> [28] "FlexVideoSideAutoLevels=B[1]*[]"                                
#> [29] "HighResVideoTopExpTime=L[60]*[]"                                
#> [30] "HighResVideoSideExpTime=L[60]*[]"                               
#> [31] "HighResVideoTopGain=L[25]*[]"                                   
#> [32] "HighResVideoSideGain=L[8]*[]"                                   
#> [33] "HighResVideoTopADCLevel=L[2]*[]"                                
#> [34] "HighResVideoSideADCLevel=L[2]*[]"                               
#> [35] "HighResVideoTopRatioR=D[1]*[]"                                  
#> [36] "HighResVideoSideRatioR=D[1]*[]"                                 
#> [37] "HighResVideoTopRatioG=D[0.8]*[]"                                
#> [38] "HighResVideoSideRatioG=D[1]*[]"                                 
#> [39] "HighResVideoTopRatioB=D[0.8]*[]"                                
#> [40] "HighResVideoSideRatioB=D[1]*[]"                                 
#> [41] "HighResVideoTopRatioGamma=D[0.9]*[]"                            
#> [42] "HighResVideoSideRatioGamma=D[1]*[]"                             
#> [43] "HighResVideoTopAutoLevels=B[1]*[]"                              
#> [44] "HighResVideoSideAutoLevels=B[1]*[]"                             
#> [45] "ActiveSensorByGUID=S[{ABB75273-9543-431a-B681-C79B533DD9E6}]*[]"
#> [46] "MeasEnv=L[0]*[]"                                                
#> [47] "SystemStateIdleZAxisMode=L[0]*[]"                               
#> [48] "SystemStateIdleMixOut4Mode=L[3]*[]"                             
#> [49] "SystemStateIdleXYAxisMode=L[0]*[]"                              
#> [50] "SystemStateIdleZAxisValue=D[0]*[m]"                             
#> [51] "KPFMModeEnabled=B[0]*[]"                                        
#> 
#> [[20]]
#>  [1] "[DataSet-0:1]"      "Version=2"          "Points=128"        
#>  [4] "Lines=128"          "Frame=Scan forward" "CurLine=127"       
#>  [7] "Dim0Name=X*"        "Dim0Unit=m"         "Dim0Range=1e-005"  
#> [10] "Dim0Min=0"          "Dim1Name=Y*"        "Dim1Unit=m"        
#> [13] "Dim1Range=1e-005"   "Dim1Min=0"          "Dim2Name=Z-Axis"   
#> [16] "Dim2Unit=m"         "Dim2Range=2e-005"   "Dim2Min=-1e-005"   
#> [19] "SaveMode=Binary"    "SaveBits=16"        "SaveSign=Signed"   
#> [22] "SaveOrder=Intel"   
#> 
#> [[21]]
#>  [1] "[DataSet-0:2]"      "Version=2"          "Points=128"        
#>  [4] "Lines=128"          "Frame=Scan forward" "CurLine=127"       
#>  [7] "Dim0Name=X*"        "Dim0Unit=m"         "Dim0Range=1e-005"  
#> [10] "Dim0Min=0"          "Dim1Name=Y*"        "Dim1Unit=m"        
#> [13] "Dim1Range=1e-005"   "Dim1Min=0"          "Dim2Name=Amplitude"
#> [16] "Dim2Unit=V"         "Dim2Range=5"        "Dim2Min=-2.5"      
#> [19] "SaveMode=Binary"    "SaveBits=16"        "SaveSign=Signed"   
#> [22] "SaveOrder=Intel"   
#> 
#> [[22]]
#>  [1] "[DataSet-1:1]"       "Version=2"           "Points=128"         
#>  [4] "Lines=128"           "Frame=Scan backward" "CurLine=127"        
#>  [7] "Dim0Name=X*"         "Dim0Unit=m"          "Dim0Range=1e-005"   
#> [10] "Dim0Min=0"           "Dim1Name=Y*"         "Dim1Unit=m"         
#> [13] "Dim1Range=1e-005"    "Dim1Min=0"           "Dim2Name=Z-Axis"    
#> [16] "Dim2Unit=m"          "Dim2Range=2e-005"    "Dim2Min=-1e-005"    
#> [19] "SaveMode=Binary"     "SaveBits=16"         "SaveSign=Signed"    
#> [22] "SaveOrder=Intel"    
#> 
#> [[23]]
#>  [1] "[DataSet-1:2]"       "Version=2"           "Points=128"         
#>  [4] "Lines=128"           "Frame=Scan backward" "CurLine=127"        
#>  [7] "Dim0Name=X*"         "Dim0Unit=m"          "Dim0Range=1e-005"   
#> [10] "Dim0Min=0"           "Dim1Name=Y*"         "Dim1Unit=m"         
#> [13] "Dim1Range=1e-005"    "Dim1Min=0"           "Dim2Name=Amplitude" 
#> [16] "Dim2Unit=V"          "Dim2Range=5"         "Dim2Min=-2.5"       
#> [19] "SaveMode=Binary"     "SaveBits=16"         "SaveSign=Signed"    
#> [22] "SaveOrder=Intel"    
#> 
#> [[24]]
#> [1] "[SetView]"           "ViewWidth=653"       "ViewHeight=527"     
#> [4] "ViewCount=4"         "View0=SetView-View0" "View1=SetView-View1"
#> [7] "View2=SetView-View2" "View3=SetView-View3"
#> 
#> [[25]]
#>  [1] "[SetView-View0]"    "ViewTyp=1"          "LineOp=2"          
#>  [4] "Channel=1"          "Frame=0"            "ZRange=-283,122"   
#>  [7] "ViewSize=256"       "ViewMode=0"         "GraphMode=1"       
#> [10] "AspectRatio=0"      "AutoZRange=1"       "ParCount=3"        
#> [13] "Par0=1,Shading,0"   "Par1=1,DeltaGain,5" "Par2=1,Smooth,0"   
#> 
#> [[26]]
#>  [1] "[SetView-View1]"      "ViewTyp=0"            "LineOp=1"            
#>  [4] "Channel=1"            "Frame=0"              "ZRange=-548,198"     
#>  [7] "ViewSize=256"         "ViewMode=0"           "GraphMode=1"         
#> [10] "AspectRatio=0"        "AutoZRange=1"         "ParCount=1"          
#> [13] "Par0=0,ReverseLine,1"
#> 
#> [[27]]
#>  [1] "[SetView-View2]"    "ViewTyp=1"          "LineOp=2"          
#>  [4] "Channel=2"          "Frame=0"            "ZRange=-168,167"   
#>  [7] "ViewSize=256"       "ViewMode=0"         "GraphMode=1"       
#> [10] "AspectRatio=0"      "AutoZRange=1"       "ParCount=3"        
#> [13] "Par0=1,Shading,0"   "Par1=1,DeltaGain,5" "Par2=1,Smooth,0"   
#> 
#> [[28]]
#>  [1] "[SetView-View3]"      "ViewTyp=0"            "LineOp=1"            
#>  [4] "Channel=2"            "Frame=0"              "ZRange=-308,375"     
#>  [7] "ViewSize=256"         "ViewMode=0"           "GraphMode=1"         
#> [10] "AspectRatio=0"        "AutoZRange=1"         "ParCount=1"          
#> [13] "Par0=0,ReverseLine,1"
#>